+++ src/Exception/NotFoundException.php
@@ -1,10 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
 namespace ProposalPage\Sdk\Exception;
 
 use Exception;
 
-class NotFoundException extends Exception
+final class NotFoundException extends Exception
 {
     public function __construct()
     {